Add DTO suffix to API objects

+++ api/specification/types.ts
@@ -1,7 +1,7 @@
-import {Int32, String} from "@airtasker/spot";
+import {Int64, String} from "@airtasker/spot";
 
 /** Unique resource identifier */
-export type Identifier = Int32;
+export type Identifier = Int64;
 
 /** [RFC 7519](https://tools.ietf.org/html/rfc7519) compliant authentication token */
 export type JWT = String;
